cloud-computing/requirements.txt
Flask==2.0.1
google-cloud-storage>=1.42.0,<2.0.0
pymysql==1.0.2
gunicorn==20.1.0
Werkzeug==2.0.2
pytz==2021.1